Add wandb-query CLI
