+++ external/wandb-query/setup.cfg
@@ -1,0 +1,11 @@
+[metadata]
+name = wandb-query
+version = 0.1.0
+description = A CLI for querying wandb. Helps avoid button-clicking.
+license = MIT
+
+[options]
+scripts = wandb-query
+install_requires =
+    wandb >= 0.12.1
+    typer >= 0.4.1
